@@ -6,26 +6,29 @@
   "license": "MIT",
   "repository": "https://github.com/probot/mistaken-pull-closer.git",
   "scripts": {
-    "lint": "stardard",
-    "start": "probot run ./index.js",
-    "test:standard": "standard",
-    "test:jest": "jest",
-    "test": "npm-run-all test:jest test:standard"
+    "build": "tsc -p tsconfig.json",
+    "dev": "nodemon --exec \"npm start\"",
+    "lint": "standard --fix",
+    "start": "probot run ./lib/index.js",
+    "test:watch": "jest --watch --notify --notifyMode=change --coverage",
+    "test": "jest && standard"
   },
   "dependencies": {
     "probot": "^7.0.1",
     "probot-config": "^0.2.0"
   },
   "devDependencies": {
-    "eslint-config-probot": ">=0.1.0",
-    "jest": "^22.4.4",
-    "localtunnel": "^1.8.3",
+    "@types/jest": "^23.3.0",
+    "@types/node": "^10.5.3",
+    "jest": "^23.4.1",
+    "nodemon": "^1.17.2",
     "npm-run-all": "^4.1.3",
     "smee-client": "^1.0.2",
-    "standard": "^10.0.3"
+    "standard": "^10.0.3",
+    "ts-jest": "^23.0.1",
+    "typescript": "^2.9.2"
   },
   "engines": {
-    "node": "^7.7",
-    "npm": "^4.0"
+    "node": ">= 8.3.0"
   }
 }
