@@ -1,17 +1,15 @@
-/* eslint-env jest */
-
-const {Application} = require('probot')
-const defaultConfig = require('../default-config')
-const startApp = require('../index')
-const pullRequestFromReleaseBranch =
-    require('./fixtures/pull-request-from-release-branch')
+import {Application, Context} from 'probot'
+import defaultConfig from '../src/default-config'
+import startApp from '../src/index'
+import * as pullRequestFromReleaseBranch from './fixtures/pull-request-from-release-branch.json'
 
 const defaultCommentBody = defaultConfig.commentBody
 
-let app
-let github
+let app: any
+let config: any
+let github: any
 
-function sendPullRequest (payload) {
+function sendPullRequest (payload: any) {
   return app.receive({
     name: 'pull_request.opened',
     event: 'pull_request',
@@ -23,23 +21,19 @@ function bareJest () {
   return jest.fn().mockReturnValue(Promise.resolve())
 }
 
-// TODO: We should be mocking probot-config out completely rather than building upon
-//       assumptions about its implementation details.
-function setConfig (config) {
+function getConfig (context: Context, fileName: string, defaultConfig?: object): Promise<any> {
   if (config) {
-    github.repos.getContent = jest.fn().mockReturnValue(Promise.resolve({
-      data: {
-        content: Buffer.from(JSON.stringify(config)).toString('base64')
-      }
-    }))
+    return Promise.resolve(Object.assign({}, defaultConfig || {}, config))
   } else {
-    let error = new Error()
-    error.code = 404
-    github.repos.getContent = jest.fn().mockReturnValue(Promise.reject(error))
+    return Promise.resolve(null)
   }
 }
 
-function setPermissionLevel (level) {
+function setConfig (tempConfig: any) {
+  config = tempConfig
+}
+
+function setPermissionLevel (level: any) {
   github.repos.reviewUserPermissionLevel =
     jest.fn().mockReturnValue(Promise.resolve({
       data: {
@@ -65,7 +59,7 @@ describe('mistaken-pull-closer', async () => {
 
     app.auth = () => Promise.resolve(github)
 
-    startApp(app)
+    startApp(app, getConfig)
   })
 
   describe('when the default configuration is used', async () => {
@@ -115,7 +109,7 @@ describe('mistaken-pull-closer', async () => {
       })
     })
 
-    describe('and a normal PR is opened', () => {
+    describe('and a normal PR is opened', async () => {
       beforeEach(async () => {
         setPermissionLevel('admin')
         await sendPullRequest(pullRequestFromReleaseBranch)
@@ -131,7 +125,7 @@ describe('mistaken-pull-closer', async () => {
     })
   })
 
-  describe('when an alternate message is configured', () => {
+  describe('when an alternate message is configured', async () => {
     it('uses the configured message', async () => {
       const testComment = 'test comment'
       setConfig({commentBody: testComment})
@@ -148,7 +142,7 @@ describe('mistaken-pull-closer', async () => {
     })
   })
 
-  describe('when an alternate label is configured', () => {
+  describe('when an alternate label is configured', async () => {
     it('creates the configured label', async () => {
       github.issues.getLabel = jest.fn().mockReturnValue(Promise.reject(new Error()))
       setConfig({labelName: 'autoclosed', labelColor: 'c0ffee'})
@@ -165,7 +159,7 @@ describe('mistaken-pull-closer', async () => {
     })
   })
 
-  describe('when addLabel is false', function () {
+  describe('when addLabel is false', async () => {
     beforeEach(async () => {
       setConfig({addLabel: false})
       setPermissionLevel('read')
